pyproject.toml
[project]
name = "climate-stable-framework"
version = "0.1.0"
description = "Climate pipeline implemented on top of the stable deployment framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==2.9.3",
    "boto3>=1.34.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "pyarrow>=15.0.0",
    "requests>=2.31.0",
]
